+++ rules/Renaming/NodeManipulator/ClassRenamer.php
@@ -28,6 +28,7 @@ use Rector\CodingStyle\Naming\ClassNaming;
 use Rector\Core\Configuration\Option;
 use Rector\Core\Configuration\Parameter\ParameterProvider;
 use Rector\Core\PhpParser\Node\BetterNodeFinder;
+use Rector\Core\Util\Hasher;
 use Rector\Naming\Naming\UseImportsResolver;
 use Rector\NodeNameResolver\NodeNameResolver;
 use Rector\NodeRemoval\NodeRemover;
@@ -450,14 +451,9 @@ final class ClassRenamer
     {
         $oldToNewClasses = $this->resolveOldToNewClassCallbacks($node, $oldToNewClasses);
 
-        // md4 is faster then md5 https://php.watch/articles/php-hash-benchmark
-        $hashingAlgorithm = 'md4';
-        if (\PHP_VERSION_ID >= 80100) {
-            // if xxh128 is available use it, as it is way faster then md4 https://php.watch/articles/php-hash-benchmark
-            $hashingAlgorithm = 'xxh128';
-        }
-
-        $cacheKey = \hash($hashingAlgorithm, \serialize($oldToNewClasses));
+        $hasher = new Hasher();
+        $serialized = \serialize($oldToNewClasses);
+        $cacheKey = $hasher->hash($serialized);
 
         if (isset($this->oldToNewTypesByCacheKey[$cacheKey])) {
             return $this->oldToNewTypesByCacheKey[$cacheKey];
